test test test

+++ triplePointApi/src/main/java/com/triple/triplePointApi/service/PointServiceImpl.java
@@ -2,8 +2,7 @@ package com.triple.triplePointApi.service;
 
 import com.triple.triplePointApi.domain.Point;
 import com.triple.triplePointApi.domain.PointHistory;
-import com.triple.triplePointApi.exception.AbnormalPointException;
-import com.triple.triplePointApi.exception.DuplicateDataException;
+import com.triple.triplePointApi.exception.*;
 import com.triple.triplePointApi.repository.PointHistoryRepositoryImpl;
 import com.triple.triplePointApi.repository.PointRepository;
 import com.triple.triplePointApi.repository.PointRepositoryImpl;
@@ -14,6 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -31,7 +31,7 @@ public class PointServiceImpl implements PointService{
     /**
      * 사용자 신규 생성시
      **/
-    public String createNewUserPoint(String userId) {
+    public String createNewUserPoint(String userId) throws DuplicateDataException {
         //중복회원이 있을 경우 에러 발생
         validateDuplicateUser(userId);
         //포인트 히스토리 생성
@@ -49,7 +49,14 @@ public class PointServiceImpl implements PointService{
     /**
      * 리뷰 최초 등록시
      **/
-    public String addReviewPoint(String userId, String placeId, List<String> photos) {
+    public String addReviewPoint(String userId, String placeId, String reviewId, List<String> photos) throws NotEnoughPointException, NoUserPointDataException, DuplicateDataException {
+        // 해당 사용자가 남긴이력이 존재하는데
+        // 해당 이력이 삭제 이력이 아닌 경우
+        Optional<PointHistory> history = pointHistoryRepositoryImpl.findMostRecentPointHistory(userId, placeId);
+        if(history.isPresent() && history.get().getDiscriminator() !="DELETE"){
+            throw new DuplicateDataException("이미 등록된 리뷰가 있습니다.");
+        }
+
         int addPoint = calculator.createAddPoint(placeId, photos);
         boolean photo = calculator.existPhoto(photos);
         boolean bonus = calculator.bonus(placeId);
@@ -58,8 +65,9 @@ public class PointServiceImpl implements PointService{
 
         //포인트 히스토리 리포지토리 호출
         PointHistory pointHistory = PointHistory.createPointHistory(
-                point, addPoint, "ADD", userId, placeId
+                point, addPoint, "ADD", userId, placeId, reviewId
                 , photo, bonus);
+        pointHistoryRepositoryImpl.create(pointHistory);
 
         return point.getUuid();
     }
@@ -67,7 +75,13 @@ public class PointServiceImpl implements PointService{
 
 
     /**리뷰 수정시**/
-    public String modReviewPoint(String userId, String placeId, List<String> photos) {
+    public String modReviewPoint(String userId, String placeId, String reviewId, List<String> photos) throws NotEnoughPointException, NoUserPointDataException, NoSuitableReviewException {
+        //수정할 리뷰에 대한 이력이 존재하지 않거나
+        //혹은 존재할경우 이미 삭제 됐을 경우
+        Optional<PointHistory> history = pointHistoryRepositoryImpl.findMostRecentPointHistoryByReviewId(reviewId);
+        if(history.isEmpty() || history.get().getDiscriminator() =="DELETE"){
+            throw new NoSuitableReviewException("수정할 리뷰내역이 존재하지 않습니다.");
+        }
 
         int modifiedPoint = calculator.modifiedPoint(userId, placeId, photos);
         boolean photo = calculator.existPhoto(photos);
@@ -78,7 +92,7 @@ public class PointServiceImpl implements PointService{
 
         //포인트 히스토리 리포지토리 호출
         PointHistory pointHistory = PointHistory.createPointHistory(
-                point, modifiedPoint, "MOD", userId, placeId
+                point, modifiedPoint, "MOD", userId, placeId, reviewId
                 , photo, bonus);
         pointHistoryRepositoryImpl.create(pointHistory);
 
@@ -92,17 +106,29 @@ public class PointServiceImpl implements PointService{
      **/
     // 해당 리뷰로 부여한 점수 삭제
     // 포인트 히스토리에서 해당 리뷰 총합계 구한뒤 이를 사용자 포인트에서 차감
-    public boolean deleteReview(String userId, String placeId) {
+    public boolean deleteReview(String userId, String placeId, String reviewId) throws AbnormalPointException, NotEnoughPointException, NoUserPointDataException, NoSuitableReviewException {
+        //수정할 리뷰에 대한 이력이 존재하지 않거나
+        //혹은 존재할경우 이미 삭제 됐을 경우
+        Optional<PointHistory> history = pointHistoryRepositoryImpl.findMostRecentPointHistoryByReviewId(reviewId);
+        if(history.isEmpty() || history.get().getDiscriminator() =="DELETE"){
+            throw new NoSuitableReviewException("삭제할 리뷰내역이 존재하지 않습니다.");
+        }
+
+
+
         int deletePoint = calculator.deletePoint(userId, placeId);
 
-        if (deletePoint >= 0) {
+        System.out.println("===========PointServiceImpl.deleteReview deletePoint : "+deletePoint);
+
+        if (deletePoint > 0) { //삭제할 리뷰가 있다는건 최소 1점은 차감된다는것.
             deletePoint *= -1; // 삭제할 포인트이므로 음수로 변경
 
             Point point = getPoint(userId, deletePoint);
 
             PointHistory pointHistory = PointHistory.createPointHistory(
-                    point, deletePoint, "DELETE", userId, placeId
+                    point, deletePoint, "DELETE", userId, placeId, reviewId
                     , false, false);
+
             pointHistoryRepositoryImpl.create(pointHistory);
 
             return true;
@@ -126,6 +152,7 @@ public class PointServiceImpl implements PointService{
         return allPoint;
     }
 
+    //회원별 포인트 조회
     public int getPointByUserId(String userId) {
         return pointRepository.getPointByUserId(userId).getPoint();
     }
@@ -133,14 +160,14 @@ public class PointServiceImpl implements PointService{
 
 
 
-    private Point getPoint(String userId, int deletePoint) {
+    private Point getPoint(String userId, int deletePoint) throws NotEnoughPointException, NoUserPointDataException {
         String modifiedPointId = pointRepository.modifyPoint(userId, deletePoint);
         Point point = em.find(Point.class, modifiedPointId);
         return point;
     }
 
     //중복 회원 검사
-    private void validateDuplicateUser(String userId) {
+    private void validateDuplicateUser(String userId) throws DuplicateDataException {
         Point pointByUserId = pointRepository.getPointByUserId(userId);
 
         if (pointByUserId != null) {
